Optimizes the search function

+++ backend/db/schemas/search.py
@@ -5,6 +5,6 @@ class Search(BaseModel):
     ingredient_names: List[str]
 
 class SearchResult(BaseModel):
-    recipe_names: dict[str,List[float]]
+    recipe_names: dict[str,List[int]]
     class Config():
         orm_mode = True
